requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.1
Pillow==8.0.1
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
